HACK for when loading the ONNX file.
I hacked the way to load the ONNX file that is exported from python.
So in MATLAB we need to modify a file in the ONNX toolbox. This file is
translateMaxPool.m . I have palced the file in the toolbox just to keep
it somewhere.
Furthermore, I have added an exported version of the ONNXmodel, so that
we dont have to export the model everytime.

This works on my machine: but we need to validate on other pcs.